修改清收程序显示问题，修改拍卖记录信息

+++ p10-casee/p10-casee-api/src/main/java/com/pig4cloud/pig/casee/entity/paifuentity/entityzxprocedure/PaiFu_STCC_PMJG_PMJG.java
@@ -25,7 +25,7 @@ public class PaiFu_STCC_PMJG_PMJG extends CommonalityData implements Serializabl
 	/**
 	 * 成交价格
 	 */
-	private Double dealPrice;
+	private BigDecimal dealPrice;
 
 	/**
 	 * 参拍人数
